feat: add initial domain logic for batches
This is the result of working through chapter 1. We
want to create the domain models `Batch` and `OrderLine`
to model the `allocations` domain. In addition, we have created
a domain service `allocate` for allocating an `Orderline`
with an available `Batch`, prioritizing batches that have
been in the warehouse the longest.